+++ instagram_private_api/endpoints/tags.py
@@ -1,5 +1,7 @@
 import json
 
+from ..compat import compat_urllib_parse
+
 
 class TagsEndpointsMixin(object):
     """For endpoints in ``/tags/``."""
@@ -11,7 +13,8 @@ class TagsEndpointsMixin(object):
         :param tag:
         :return:
         """
-        endpoint = 'tags/{tag!s}/info/'.format(**{'tag': tag})
+        endpoint = 'tags/{tag!s}/info/'.format(
+            **{'tag': compat_urllib_parse.quote(tag.encode('utf8'))})
         res = self._call_api(endpoint)
         return res
 
@@ -22,7 +25,8 @@ class TagsEndpointsMixin(object):
         :param tag:
         :return:
         """
-        endpoint = 'tags/{tag!s}/related/'.format(**{'tag': tag})
+        endpoint = 'tags/{tag!s}/related/'.format(
+            **{'tag': compat_urllib_parse.quote(tag.encode('utf8'))})
         query = {
             'visited': json.dumps([{'id': tag, 'type': 'hashtag'}], separators=(',', ':')),
             'related_types': json.dumps(['hashtag', 'location'], separators=(',', ':'))}
@@ -45,3 +49,39 @@ class TagsEndpointsMixin(object):
         query.update(kwargs)
         res = self._call_api('tags/search/', query=query)
         return res
+
+    def tags_user_following(self, user_id):
+        """
+        Get tags a user is following
+
+        :param user_id:
+        :return:
+        """
+        endpoint = 'users/{user_id!s}/following_tags_info/'.format(user_id=user_id)
+        return self._call_api(endpoint)
+
+    def tag_follow_suggestions(self):
+        """Get suggestions for tags to follow"""
+        return self._call_api('tags/suggested/')
+
+    def tag_follow(self, tag):
+        """
+        Follow a tag
+
+        :param tag:
+        :return:
+        """
+        endpoint = 'tags/follow/{hashtag!s}/'.format(
+            hashtag=compat_urllib_parse.quote(tag.encode('utf-8')))
+        return self._call_api(endpoint, params=self.authenticated_params)
+
+    def tag_unfollow(self, tag):
+        """
+        Unfollow a tag
+
+        :param tag:
+        :return:
+        """
+        endpoint = 'tags/unfollow/{hashtag!s}/'.format(
+            hashtag=compat_urllib_parse.quote(tag.encode('utf-8')))
+        return self._call_api(endpoint, params=self.authenticated_params)